changes to helper db methods
